pyproject.toml
[project]
name = "llm-games"
version = "0.1.0"
description = "AutoGen quickstart example with countdown agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "autogen-core>=0.1.0",
    "autogen-ext>=0.7.4",
    "autogen-agentchat>=0.7.4",
    "asyncio-mqtt>=0.16.0",
    "pydantic>=2.0.0",
    "openai>=1.100.0",
    "tiktoken>=0.11.0",
    "python-dotenv>=1.0.0",
    "autogen>=0.9.9",
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "langchain>=0.3.27",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
